project/rosa_agent/ReplicateRosaMilestone/requirements.txt
langchain
langchain-openai
langchain-community
python-dotenv
nats-py
streamlit
